webapp/python/requirements.txt
Flask
SQLAlchemy
mysql-connector-python